move stash-merritt from a gem structure to be a core part of the Repo module
