+++ lib/datacite/mapping/datacite_xml_factory.rb
