+++ lib/stash/repo/builders.rb
@@ -1,5 +1,5 @@
 module Stash
-  module Merritt
+  module Repo
     module Builders
       Dir.glob(File.expand_path('builders/*.rb', __dir__)).each(&method(:require))
     end
