+++ lib/stash/repo/data_one_manifest_builder.rb
@@ -1,0 +1,29 @@
+require 'ostruct'
+require 'merritt'
+
+module Stash
+  module Repo
+    class DataONEManifestBuilder < FileBuilder
+      attr_reader :uploads
+      
+      # @param uploads [Array[StashEngine::DataFile]] a list of file uploads
+      def initialize(uploads)
+        super(file_name: 'mrt-dataone-manifest.txt')
+        @uploads = uploads
+      end
+      
+      def mime_type
+        MIME::Types['text/plain'].first
+      end
+      
+      def contents
+        files = uploads.map do |upload|
+          OpenStruct.new(file_name: upload.upload_file_name, mime_type: upload.upload_content_type)
+        end
+        manifest = ::Merritt::Manifest::DataONE.new(files: files)
+        manifest.write_to_string
+      end
+    end
+  end
+end
+
