+++ lib/stash/repo/schemas/datacite/include/datacite-contributorType-v4.xsd
