+++ lib/stash/repo/schemas/datacite/include/datacite-descriptionType-v4.xsd
