+++ lib/stash/repo/schemas/datacite/include/datacite-nameType-v4.xsd
