+++ lib/stash/repo/schemas/datacite/include/datacite-relatedIdentifierType-v4.xsd
