+++ lib/stash/repo/schemas/datacite/include/datacite-resourceType-v4.xsd
