+++ lib/stash/repo/schemas/datacite/include/xml.xsd
