+++ lib/stash/repo/schemas/datacite/metadata.xsd
