+++ lib/stash/repo/schemas/stash-wrapper.xsd
