--- stash/stash-merritt/lib/stash.rb
@@ -1,5 +1,0 @@
-require 'logger'
-
-module Stash
-  Dir.glob(File.expand_path('stash/*.rb', __dir__)).each(&method(:require))
-end
