--- stash/stash-merritt/lib/stash/merritt/module_info.rb
@@ -1,12 +1,0 @@
-module Stash
-  module Merritt
-    # The name of this gem
-    NAME = 'stash-merritt'.freeze
-
-    # The version of this gem
-    VERSION = '0.2.5'.freeze
-
-    # The copyright notice for this gem
-    COPYRIGHT = 'Copyright (c) 2017 The Regents of the University of California'.freeze
-  end
-end
